Refactor namespace and file structure for TransactionDraft
Moved TransactionDraft-related files from Common to a more specific Wallet.Operate.Send namespace and directory for improved organization. Updated references, data context, and validation rules to align with the new structure.
