+++ src/Angor/Avalonia/AngorApp/Sections/Wallet/Operate/Send/AddressAndAmount/AddressAndAmountView.axaml.cs
@@ -1,4 +1,4 @@
-namespace AngorApp.Sections.Wallet.Operate.Send;
+namespace AngorApp.Sections.Wallet.Operate.Send.AddressAndAmount;
 
 public partial class AddressAndAmountView : UserControl
 {
