+++ src/Angor/Avalonia/AngorApp/Sections/Wallet/Operate/Send/AddressAndAmount/AddressAndAmountViewModel.cs
@@ -2,7 +2,7 @@ using ReactiveUI.SourceGenerators;
 using ReactiveUI.Validation.Extensions;
 using ReactiveUI.Validation.Helpers;
 
-namespace AngorApp.Sections.Wallet.Operate.Send;
+namespace AngorApp.Sections.Wallet.Operate.Send.AddressAndAmount;
 
 public partial class AddressAndAmountViewModel : ReactiveValidationObject, IAddressAndAmountViewModel
 {
@@ -13,10 +13,10 @@ public partial class AddressAndAmountViewModel : ReactiveValidationObject, IAddr
     {
         walletBalanceHelper = wallet.WhenAnyValue(wallet1 => wallet1.Balance).ToProperty(this, model => model.WalletBalance);
         
-        this.ValidationRule(x => x.Amount, x => x is null or > 0, _ => "Amount must be greater than zero");
-        this.ValidationRule(x => x.Amount, x => x is not null, _ => "Please, specify an amount");
-        var isValidAmount = this.WhenAnyValue(x => x.Amount, x => x.WalletBalance, (a, b) => a is null || a <= b);
-        this.ValidationRule(x => x.Amount, isValidAmount, "Amount exceeds balance");
+        this.ValidationRule<AddressAndAmountViewModel, long?>(x => x.Amount, x => x is null or > 0, _ => "Amount must be greater than zero");
+        this.ValidationRule<AddressAndAmountViewModel, long?>(x => x.Amount, x => x is not null, _ => "Please, specify an amount");
+        var isValidAmount = this.WhenAnyValue<AddressAndAmountViewModel, bool, long?, long>(x => x.Amount, x => x.WalletBalance, (a, b) => a is null || a <= b);
+        this.ValidationRule<AddressAndAmountViewModel, long?>(x => x.Amount, isValidAmount, "Amount exceeds balance");
 
         this.ValidationRule<AddressAndAmountViewModel, string>(x => x.Address, x => !string.IsNullOrWhiteSpace(x), _ => "Please, specify an address");
         this.ValidationRule<AddressAndAmountViewModel, string>(x => x.Address, x => string.IsNullOrWhiteSpace(x) || wallet.IsAddressValid(x).IsSuccess, message => wallet.IsAddressValid(message).Error);
