+++ src/Angor/Avalonia/AngorApp/Sections/Wallet/Operate/Send/TransactionDraft/ITransactionDraftViewModel.cs
@@ -1,7 +1,7 @@
 using Angor.Contexts.Wallet.Domain;
 using Zafiro.Avalonia.Controls.Wizards.Builder;
 
-namespace AngorApp.UI.Controls.Common.TransactionDraft;
+namespace AngorApp.Sections.Wallet.Operate.Send.TransactionDraft;
 
 public interface ITransactionDraftViewModel : IStep
 {
