+++ src/Angor/Avalonia/AngorApp/Sections/Wallet/Operate/Send/TransactionDraft/TransactionDraftView.axaml
@@ -4,15 +4,15 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:fa="clr-namespace:FluentAvalonia.UI.Controls;assembly=FluentAvalonia"
              xmlns:controls="clr-namespace:AngorApp.UI.Controls"
-             xmlns:transactionPreview="clr-namespace:AngorApp.UI.Controls.Common.TransactionDraft"
+             xmlns:transactionDraft="clr-namespace:AngorApp.Sections.Wallet.Operate.Send.TransactionDraft"
              mc:Ignorable="d"
              d:DesignWidth="600"
              d:DesignHeight="400"
-             x:Class="AngorApp.UI.Controls.Common.TransactionDraft.TransactionDraftView"
-             x:DataType="transactionPreview:ITransactionDraftViewModel">
+             x:Class="AngorApp.Sections.Wallet.Operate.Send.TransactionDraft.TransactionDraftView"
+             x:DataType="transactionDraft:ITransactionDraftViewModel">
 
     <Design.DataContext>
-        <transactionPreview:TransactionDraftViewModelDesign />
+        <transactionDraft:TransactionDraftViewModelDesign />
     </Design.DataContext>
 
     <UserControl.Styles>
@@ -26,7 +26,7 @@
     </UserControl.Styles>
 
     <StackPanel Spacing="8">
-        <transactionPreview:Summary />
+        <transactionDraft:Summary />
         <Separator />
         <Button HorizontalAlignment="Center" Margin="0 10" Command="{Binding Confirm}">
             <Button.Content>
