+++ src/Angor/Avalonia/AngorApp/Sections/Wallet/Operate/Send/TransactionDraft/TransactionDraftView.axaml.cs
@@ -1,4 +1,4 @@
-namespace AngorApp.UI.Controls.Common.TransactionDraft;
+namespace AngorApp.Sections.Wallet.Operate.Send.TransactionDraft;
 
 public partial class TransactionDraftView : UserControl
 {
